add examples folder; lint the examples
